:sparkles: move queries to module
